Legends Website changed so I fixed it and let it create backups.
Other errors I missed were also fixed, major changes in a few classes.

+++ src/main/java/com/github/egubot/build/OnlineDataManager.java
@@ -1,10 +1,12 @@
 package com.github.egubot.build;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -31,15 +33,27 @@ public class OnlineDataManager {
 
 	private List<String> data = new ArrayList<>(0);
 	private int lockedDataEndIndex;
+	private String lastUpdateDate = null;
 	private String dataName;
 
-	public OnlineDataManager(DiscordApi api, String storageKey, String resourcePath, String dataName) throws Exception {
+	public OnlineDataManager(DiscordApi api, String storageKey, String dataName, InputStream localInput,
+			boolean verbose) throws Exception {
+		this.api = api;
+		this.storageKey = storageKey;
+		this.storageMsgID = KeyManager.getID(storageKey);
+		this.dataName = dataName;
+		this.localInputStream = localInput;
+		initialise(verbose);
+	}
+
+	public OnlineDataManager(DiscordApi api, String storageKey, String resourcePath, String dataName, boolean verbose)
+			throws Exception {
 		this.api = api;
 		this.storageKey = storageKey;
 		this.storageMsgID = KeyManager.getID(storageKey);
 		this.dataName = dataName;
 		findLocalInput(resourcePath);
-		initialise();
+		initialise(verbose);
 	}
 
 	private void findLocalInput(String resourcePath) {
@@ -49,31 +63,34 @@ public class OnlineDataManager {
 				localInputStream = new FileInputStream(new File(resourcePath));
 			}
 		} catch (Exception e) {
-			System.err.println("\nWarning: no local " + dataName + " data. Expected " + resourcePath +" to be present.");
+			System.err
+					.println("\nWarning: no local " + dataName + " data. Expected " + resourcePath + " to be present.");
 			localInputStream = null;
 		}
 	}
 
-	private void initialise() throws Exception {
+	private void initialise(boolean verbose) throws Exception {
 		try {
 			storageMessage = api.getMessageById(storageMsgID, api.getTextChannelById(storageChannelID).get()).get();
 
 		} catch (Exception e) {
-			checkStorageMessage();
+			checkStorageMessage(verbose);
 		}
 
-		getOnlineData(true);
+		getOnlineData(verbose);
 	}
 
-	private void checkStorageMessage() throws Exception {
+	private void checkStorageMessage(boolean verbose) throws Exception {
 		if (!storageChannelID.equals("-1")) {
 			uploadLocalData(true);
 			KeyManager.updateKeys(storageKey, storageMsgID, KeyManager.idsFileName);
 			storageMsgID = KeyManager.getID(storageKey);
 			try {
 				storageMessage = api.getMessageById(storageMsgID, api.getTextChannelById(storageChannelID).get()).get();
+				if(verbose)
 				System.out.println("\nNew " + dataName + " message was created.");
 			} catch (Exception e) {
+				if(verbose)
 				System.err.println("\nFailed to create new " + dataName + " message.");
 			}
 		}
@@ -98,11 +115,11 @@ public class OnlineDataManager {
 
 			newID = api.getTextChannelById(storageChannelID).get()
 					.sendMessage(getInputStream(), dataName.replace(" ", "_") + ".txt").join().getIdAsString();
-			
+
 			try {
 				storageMessage.edit(newID).join();
 			} catch (Exception e) {
-				System.err.println("\nOnline data ID failed to update");
+				storageMsgID=newID;
 			}
 
 			getInputStream().close();
@@ -143,26 +160,30 @@ public class OnlineDataManager {
 			Message newMessage = api.getMessageById(newID, api.getTextChannelById(storageChannelID).get()).join();
 			setInputStream(newMessage.getAttachments().get(0).asInputStream());
 
+			// Date
+			date = newMessage.getCreationTimestamp().toString().split("[Tz.]");
+			lastUpdateDate = date[0] + ", " + date[1].substring(0, date[1].length() - 3);
+
 			if (verbose) {
-				date = newMessage.getCreationTimestamp().toString().split("[Tz.]");
-				System.out.println("\n" + dataName + " data successfully loaded!\nDate of last update: " + date[0]
-						+ ", " + date[1].substring(0, date[1].length() - 3));
+				System.out.println("\n" + dataName + " data successfully loaded!\nDate of last update: " + lastUpdateDate);
 			}
 
-			Scanner read = new Scanner(getInputStream());
-			String st;
+			try (BufferedReader br = new BufferedReader(new InputStreamReader(getInputStream()))) {
+				String st;
 
-			data.clear();
-			while (read.hasNextLine()) {
-				st = read.nextLine().trim().replace("\n", "");
-				if (!st.equals(""))
-					data.add(st);
+				data.clear();
+				while ((st = br.readLine()) != null) {
+					st = st.trim().replace("\n", "").replace("é", "e");
+					if (!st.equals(""))
+						data.add(st);
+				}
+
 			}
-			read.close();
-
 		} catch (IOException | NullPointerException e) {
-			System.err.println("\n" + dataName + " data failed to load, internal backup used.");
-
+			if (verbose) {
+				System.err.println("\n" + dataName + " data failed to load, internal backup used.");
+			}
+		
 			setInputStream(localInputStream);
 		}
 	}
@@ -177,10 +198,16 @@ public class OnlineDataManager {
 			setInputStream(new FileInputStream(tempDataFile));
 
 			uploadLocalData(false);
-			if (!tempDataFile.delete())
-				e.sendMessage("Couldn't update <:sad:1020780174901522442>");
-			else
-				e.sendMessage("Updated <:legudrink:804071006956159008>");
+
+			if (e != null) {
+				if (!tempDataFile.delete())
+					e.sendMessage("Couldn't update <:sad:1020780174901522442>");
+				else
+					e.sendMessage("Updated <:legudrink:804071006956159008>");
+			} else {
+				if (!tempDataFile.delete())
+					System.out.println("Couldn't update");
+			}
 		} catch (Exception e1) {
 			System.err.println("Failed to write and upload data.");
 		}
@@ -216,7 +243,7 @@ public class OnlineDataManager {
 		this.lockedDataEndIndex = lockedDataEndIndex;
 	}
 
-	List<String> getData() {
+	public List<String> getData() {
 		return data;
 	}
 
@@ -230,5 +257,13 @@ public class OnlineDataManager {
 
 	public void setLocalInputStream(InputStream localInputStream) {
 		this.localInputStream = localInputStream;
+	}
+
+	public String getLastUpdateDate() {
+		return lastUpdateDate;
+	}
+
+	public void setLastUpdateDate(String lastUpdateDate) {
+		this.lastUpdateDate = lastUpdateDate;
 	}
 }
